Move task link validator methods
